backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.2
defusedxml==0.7.1
Django==3.2.3
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
gunicorn==20.1.0 
python-dotenv==1.1.0